+++ ExpandWorldPrefabs/InfoManager.cs
@@ -2,8 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using ExpandWorldData;
-using HarmonyLib;
-using UnityEngine;
 
 namespace ExpandWorld.Prefab;