+++ publish/CHANGELOG.md
@@ -1,3 +1,6 @@
+- v1.8
+  - Adds a new field `drops` to spawn drops.
+
 - v1.7
   - Adds support for checking data of the player who triggered armor stands, cooking stations, item stands or obliterators.
   - Adds a new field `delay` to delay spawns and swaps.